Async addition I kinda changed like, a lot, which wasn't my first intention, but that's how it turned out. I moved some files, added async, cleaned a lot of stuff up. However, it should all be still backwards compatible.
