+++ MANIFEST
@@ -1,4 +1,3 @@
 # file GENERATED by distutils, do NOT edit
-ipgetter.py
 setup.cfg
 setup.py
